+++ FE/src/Header/Header.jsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
 import { Navbar, Menu, Icons, MenuBox, MainMenu, SubMenuList, SubMenu } from './Header.style.js';
-import { ReactComponent as Logo } from './assets/logo.svg';
-import { ReactComponent as Search } from './assets/search.svg';
-import { ReactComponent as Login } from './assets/login.svg';
-import { ReactComponent as Cart } from './assets/cart.svg';
+import { ReactComponent as Logo } from '../assets/logo.svg';
+import { ReactComponent as Search } from '../assets/search.svg';
+import { ReactComponent as Login } from '../assets/login.svg';
+import { ReactComponent as Cart } from '../assets/cart.svg';
 
 const Header = () => {
   const [isOpen, setIsOpen] = useState(false);
